+++ src/signatures.rs
@@ -555,6 +555,10 @@ signatures! {
     format = AdaptiveMultiRate
     value = b"#!AMR"
 
+    format = AdvancedStreamRedirector
+    value = b"<ASX "
+    value = b"<asx "
+
     format = EmbeddedOpentype
     value = b"\x00\x00\x01" offset = 8, b"\x4C\x50" offset = 34
     value = b"\x01\x00\x02" offset = 8, b"\x4C\x50" offset = 34